Refactor and modularize codebase
